Tidy some code

+++ src/bf/linker.cpp
@@ -5,14 +5,14 @@ namespace bf
 {
 bool Brainfuck::link()
 {
-	std::vector<uint16_t> jumps;
+	std::vector<int> jumps;
 	for (size_t i = 0; i < program.size(); ++i)
 	{
 		switch (program[i].opcode)
 		{
 		case bfLoopBegin:
 			program[i].opcode = bfJmpZero;
-			jumps.push_back(i);
+			jumps.push_back(static_cast<int>(i));
 			break;
 
 		case bfLoopEnd:
@@ -23,7 +23,7 @@ bool Brainfuck::link()
 			}
 			program[i].opcode = bfJmpNotZero;
 
-			program[jumps.back()].args[0] = i + 1;
+			program[static_cast<size_t>(jumps.back())].args[0] = static_cast<int>(i + 1);
 			program[i].args[0] = jumps.back() + 1;
 			jumps.pop_back();
 			break;
